Make like things more alike
Imagine that 99 Bottles was a production app instead of a fixed song.
In this case we would likely not stop at Shameless Green.
The most obvious problem with this code is the duplication
of strings.  Removing string duplication lets us practice
a very specific refactoring technique.

This technique, in the micro, has three repeated steps, and
in the macro, has a pattern of completing _horizontal_
refactorings before getting distracted by _vertical_
refactorings (the meaning of these terms will become
clear) (I hope).

The steps in the mico technique are
1. compile new code
2. execute new code
3. use the result of new code.

Sometimes steps 2 and 3 are combined, but step 1 always happens
independently.

The attack plan here is to pick the two strings that are the
most alike, and to make them *more* alike by the smallest
steps possible.

The number > 2 case is almost just like the number > 1 case
so I will start there.  If they sent a message to get
bottle/bottles they would be just alike.

The message will be #containers.  Adding a #containers method
that returns _bottles_, the most common case, without
actually sending the #containers method, is step1
of the micro refactoring.