Return to task: Add 6-packs.
1. Is this open/closed to 6-packs?  No.
2. Do I know how to make it open closed to 6-packs? No, so
3. Fix the most egregious code smell.

There are several code smells here, all pointing to the same problem.
The most obvious smell is that many methods take the same argument,
and in those methods, the code depends more on the argument than
on the rest of the class.

Look at the methods that take _number_.  This pattern means that
_number_ represents a concept that should be a full-blown object,
and that all of the methods that take _number_ should be implemented
in that new object.

This commit is the first step along the path of separating _number_ into
another object.  We choose Verse as the name for the new object, and,
using the Katrina refactoring techniques, begin by copying (not cutting),
the methods into the new class.