Proceed with removing all the _number_ arguments from methods in Verse.
Verse now contains _number_ in @number and uses #number to access it.
This 'Extract Class' refactoring requires that we send this message to get _number'
instead of passing it as an argument.

The technique for removing arguments while staying green was just used to remove the
_number_ argument from Verse#verse.  It's:
1. Add a default to the method definition.
2. Remove the arg from the sender.
3. Remove the arg/default from the method definition.

Let's use this pattern to remove the remaining _number_ arguments from the
methods in Verse.