Container step 2: Remove third argument sent.  Uh-oh, now I'm RED.
This clearly won't work.  The refactoring says I should be able to change #container
so that it does not take an argument.  The container method should depend only on
_number_, i.e., verse number, i.e., the verse this instance of Verse represents.  Here,
however, this invocation of #container isn't for the verse I'm on, it's for _number_ - 1.

I should be able to remove the arguments from all the methods in Verse and the fact that
I cannot is telling me that I have missed a concept.  What concept am I missing?