Circle back to task: Add 6-packs.
1. Is this open/closed to 6-packs? No.
2. Do you know how to make it open/closed? No, so
3. Fix the next code smell.

The first code smell to deal with is always Single Responsibility.
This code isn't terrible but we don't know what to do to make it open/closed, and we
definitely do not want to add to the if statements.

Ask the question 'If I were going to divide this class into two, where would I split it?'

Use shape, responsibility, public/private.  These criteria all suggest moving the private
methods that contain only 'if' statements into another object.

Take a close look at these methods.  They have a lot in common.
1. They contain only 'if' statements, no other code.
2. The values they return are the smallest, most atomic thing possible.  The name of the
    method states the concept, and the return value of the branch of the if statement is
    tiny, atomic answer.

These methods got shaped this way because of the refactoring technique used to dry
out the strings (back in bottles_shameless_to_dry) and our insistence that we follow the
'Extract Class' refactoring all the way and remove all arguments from methods (in
dry_to_verse).  There are a number of equally defensible versions of 'shameless green'
but, when using these refactoring techniques, all code converges here.  We're on the
stable path, and we're seeing stable intermediate solutions.

Therefore, the next step is to do another Extract Class refactoring, this time moving
all the methods containing 'if' statements into a separate class.  There are many bad
names for this class...VerseVariant is my current favorite.