Start replacing 'if' statement branches with objects.
VerseVariant has some very specific qualifies.
1. All methods contain only 'if' statements.
2. The 'if' statements all switch on _number_, which represents the exact same thing
in every case (the starting number of the verse I'm on).
3. Each 'if' statement branch returns the smallest, most atomic thing that
represents the value of the concept (method name) for that branch (value of number).

Now that we're here, we can begin following a recipe to replace the conditionals with objects.
I'll do this refactoring from top to bottom, walking through the methods one at a time.

Create a class to represent the first value we're switching on. VerseVariant#initial_container
starts with 'f number == 1', so create a class to represent VerseVariant 1. I'll call it...
VerseVariant1.

In that class, create an #initial_container method and have it just return the value for variant 1.