Create a little VerseVariant factory.
The #variant method will return something that will correctly play the variant role for
that number.

Before this transition I always knew
1) the value I was switching one, and
2) what to do when I switched.

These two responsibilities are beginning to be separated into different objects.   The factory
will be responsible for choosing the correct object based on the value of number and the
variant object chosen will just do the right thing.