Implement 6-packs.
The current arrangement of code is open/closed to 6-packs...but....
This solution is just annoying.  All of the patterns we followed worked, i.e., the
refactoring techniques are correct, and the recipe we followed to turn the conditionals
into objects is correct, but this end point has code smells.

It's annoying to have to implement two variants to make 6-packs work.  The repeating
prefix/suffix that we see here is a sign that we're missing an object, and suggests that we
have missed a concept.

We introduced the repeating prefix/suffix during the first Extract Class refactoring, where
we were forced to introduce these concepts in order to remove the _number_ parameter
from the methods.  I'll end this branch here, and start another back at the point, to see
what it would take to avoid the suffix/prefix problem.